Fikser bugs hvor sammendrag ikke kan slettes og hvor endret verdi består ved lukking av modal
